Reduced the state of repeat_first.py and repeat_previous.py, and removed the is_start/can_act flag as they are not needed.
Before the state was composed of the full deck and a time/position pointer. This made the state space huge and the MDP deterministic: the only randomness was the initial shuffling of the deck.

Now the state space is composed of the first / last k cards and the ratio of remaining cards for each suit, making the state space smaller but stochastic.